Adding progska.h header file
Compilation on my Macbook gave the following error:

progska/_progska.c:84:18: error: implicit declaration of function 'main' is invalid in C99 [-Werror,-Wimplicit-function-declaration]
    marcresult = main(num_mainargs, mainargs);

Simple fix was to create progska.h header and include in _progska.c
